Update README.md
If you don't wanna approve te request that's okay I don't plan to just 'edit' files but fully replace them with my own code and thus far all I am planning to do is add a more 'day-z'ish' style of gunplay where regardless if it's a ricochet or a 22 and it hits flesh you'll need to patch the bleed along with adding a new aiming system to replace the "point aim" which I call "intuitive aim". (using a combo of deadzone and ray cast tracing the nearest enemy within 20 feed will simply tilt the axis of the Z position/muzzle in the direction of the enemy with a very high likliehood of successfully hitting them whilst they are 10-15 feet away tops and after that have it start slowly and grow exponenitially higher in dispersion unless a cooldown of 3000ms is given between each shot and aiming would simply raise the sights near the eye ball but the disengage active(intuition aim) would then make it go back to your default setting.